+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [Unreleased]
 
+## [0.2.0] - 2025-10-02
+
+- Remove dependency on `city-state`; vendor data and add internal loader
+- Add MaxMind data build support and `rake data:update`
+- Add `rubyzip` dependency for data updates
+
 ## [0.1.0] - 2025-09-30
 
 - Initial release